+++ RELEASE_NOTES.rst
@@ -4,6 +4,25 @@ Next release
 All changes
 -----------
 
+
+v3.1.0 (2020-08-28)
+===================
+
+All changes
+-----------
+
+ixmp v3.1.0 coincides with message_ix v3.1.0.
+
+- :pull:`345`: Fix a bug in :meth:`.read_excel` when parameter data is spread across multiple sheets.
+- :pull:`363`: Expand documentation and revise installation instructions.
+- :pull:`362`: Raise Python exceptions from :class:`.JDBCBackend`.
+- :pull:`354`: Add :meth:`Scenario.items`, :func:`.utils.diff`, and allow using filters in CLI command ``ixmp export``.
+- :pull:`353`: Add functionality for storing ‘meta’ (annotations of model names, scenario names, versions, and some combinations thereof).
+
+  - Add :meth:`.Backend.add_model_name`, :meth:`~.Backend.add_scenario_name`, :meth:`~.Backend.get_model_names`, :meth:`~.Backend.get_scenario_names`, :meth:`~.Backend.get_meta`, :meth:`~.Backend.set_meta`, :meth:`~.Backend.remove_meta`.
+  - Allow these to be called from :class:`.Platform` instances.
+  - Remove :meth:`.Scenario.delete_meta`.
+
 - :pull:`349`: Avoid modifying indexers dictionary in :meth:`.AttrSeries.sel`.
 - :pull:`343`: Add region/unit parameters to :meth:`.Platform.export_timeseries_data`.
 - :pull:`347`: Preserve dtypes of index columns in :func:`.data_for_quantity`.